requirements.txt
brotlipy==0.7.0
certifi==2021.10.8
cffi 
charset-normalizer 
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
pycosat==0.6.3
python-dateutil==2.8.2
pytz==2022.6
pywin32==302
SQLAlchemy==1.4.44
Werkzeug==2.2.2
wincertstore==0.2
xlrd==2.0.1
zipp==3.11.0
